+++ app/controllers/bookings.rb
@@ -1,0 +1,10 @@
+
+class Makersbnb < Sinatra::Base
+
+  get '/bookings' do
+  end
+
+  post '/bookings' do
+  end
+
+end
